requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1